refactored saveToLocalStorage to customHook, but no function change
